+++ __tests__/rpcProvider.test.ts
@@ -2,9 +2,9 @@ import { getStarkKey, utils } from '@scure/starknet';
 
 import {
   Account,
+  Block,
   CallData,
   Contract,
-  GetBlockResponse,
   RPC,
   TransactionExecutionStatus,
   stark,
@@ -77,12 +77,6 @@ describeIfRpc('RPCProvider', () => {
     expect(typeof spec).toBe('string');
   });
 
-  test('getCode - not implemented', async () => {
-    expect(
-      rpcProvider.getCode('0x058d97f7d76e78f44905cc30cb65b91ea49a4b908a76703c54197bca90f81773')
-    ).rejects.toThrow();
-  });
-
   describe('Test Estimate message fee', () => {
     const L1_ADDRESS = '0x8359E4B0152ed5A731162D3c7B0D8D56edB165A0';
     let l1l2ContractAddress: string;
@@ -113,8 +107,8 @@ describeIfRpc('RPCProvider', () => {
 
   describe('waitForTransaction', () => {
     const receipt = {};
-    const transactionStatusSpy = jest.spyOn(rpcProvider as any, 'getTransactionStatus');
-    const transactionReceiptSpy = jest.spyOn(rpcProvider as any, 'getTransactionReceipt');
+    const transactionStatusSpy = jest.spyOn(rpcProvider.channel as any, 'getTransactionStatus');
+    const transactionReceiptSpy = jest.spyOn(rpcProvider.channel as any, 'getTransactionReceipt');
 
     const generateOptions = (o: waitForTransactionOptions) => ({ retryInterval: 10, ...o });
     const generateTransactionStatus = (
@@ -172,7 +166,7 @@ describeIfRpc('RPCProvider', () => {
   });
 
   describe('RPC methods', () => {
-    let latestBlock: GetBlockResponse;
+    let latestBlock: Block;
 
     beforeAll(async () => {
       latestBlock = await provider.getBlock('latest');