+++ __tests__/utils/stark.test.ts
@@ -1,4 +1,5 @@
-import { CallData, EstimateFeeResponse, RawArgs, json, stark } from '../../src';
+import { CallData, RawArgs, UniversalDetails, json, stark } from '../../src';
+import { EDataAvailabilityMode, FeeEstimate } from '../../src/types/api';
 import { toBigInt, toHex } from '../../src/utils/num';
 import { compiledOpenZeppelinAccount } from '../config/fixtures';
 
@@ -71,14 +72,36 @@ describe('stark', () => {
   });
 
   test('estimateFeeToBounds', () => {
-    const estimateFeeResponse: EstimateFeeResponse = {
-      gas_consumed: 100n,
-      gas_price: 10n,
-      overall_fee: 1000n,
+    const estimateFeeResponse: FeeEstimate = {
+      gas_consumed: '100',
+      gas_price: '10',
+      overall_fee: '1000',
+      unit: 'FRI',
     };
     expect(stark.estimateFeeToBounds(estimateFeeResponse)).toStrictEqual({
       l2_gas: { max_amount: '0x0', max_price_per_unit: '0x0' },
       l1_gas: { max_amount: '0x6e', max_price_per_unit: '0xf' },
     });
+  });
+
+  test('v3Details', () => {
+    const setValues = (o: {}, v: any) => Object.fromEntries(Object.keys(o).map((k) => [k, v]));
+
+    const details: UniversalDetails = {
+      tip: 99n,
+      paymasterData: [99n, 99n],
+      accountDeploymentData: [99n, 99n],
+      nonceDataAvailabilityMode: EDataAvailabilityMode.L2,
+      feeDataAvailabilityMode: EDataAvailabilityMode.L2,
+      resourceBounds: {
+        l1_gas: { max_amount: '0x99', max_price_per_unit: '0x99' },
+        l2_gas: { max_amount: '0x99', max_price_per_unit: '0x99' },
+      },
+    };
+    const detailsUndefined = setValues(details, undefined);
+    const detailsAnything = setValues(details, expect.anything());
+
+    expect(stark.v3Details(details)).toMatchObject(details);
+    expect(stark.v3Details(detailsUndefined)).toEqual(expect.objectContaining(detailsAnything));
   });
 });