+++ src/types/index.ts
@@ -8,13 +8,3 @@ export * from './typedData';
 export * from './cairoEnum';
 
 export * as RPC from './api';
-export {
-  CallL1Handler,
-  DeployedContractItem,
-  ExecutionResources,
-  FunctionInvocation,
-  GetContractAddressesResponse,
-  GetTransactionStatusResponse,
-  SequencerIdentifier,
-} from './api/sequencer';
-export * as Sequencer from './api/sequencer';